test(CNTRLPLANE-3306): migrate keycloak api code to separate file
This change moves keycloak api code from `external_oidc.go` to
its own file, `keycloak.go` . Also adds check for admin token
expiry + refreshes if needed.
